+++ src/main/java/uk/gov/register/presentation/config/ZookeeperConfiguration.java
@@ -1,4 +1,4 @@
-package uk.gov.register.presentation;
+package uk.gov.register.presentation.config;
 
 public interface ZookeeperConfiguration {
     String getZookeeperServer();
