+++ src/main/java/uk/gov/register/presentation/dao/RecentEntryIndexQueryDAO.java
@@ -1,9 +1,10 @@
-package uk.gov.register.presentation;
+package uk.gov.register.presentation.dao;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import org.skife.jdbi.v2.sqlobject.Bind;
 import org.skife.jdbi.v2.sqlobject.SqlQuery;
 import org.skife.jdbi.v2.sqlobject.customizers.RegisterMapper;
+import uk.gov.register.presentation.mapper.JsonNodeMapper;
 
 import java.util.List;
